test: 생성시 검증관련 테스트 옮김
